fix the broken specs and remove unused files

+++ api/app/dtos/organization/project_versions/compact_dto.rb
@@ -2,6 +2,6 @@ class Organization::ProjectVersions::CompactDto < OpenApiDto
   field "id", :integer
   field "number", :integer
   field "created_at", :timestamp
-  field "retention_guarantee_rate", :integer
+  field "retention_guarantee_rate", :decimal
   field "total_amount", :decimal
 end
